[#3] Refactor: common 재사용 컴포넌트들 main 화면에서 사용
