+++ src/components/product/ProductItem.jsx
@@ -1,3 +1,7 @@
+import ProductName from '../common/ProductName';
+import ProductNumber from '../common/ProductNumber';
+import ProductPrice from '../common/ProductPrice';
+import ThumbNailImage from '../common/ThumbNailImage';
 import { useProductActionContext } from './../../contexts/ProductContext';
 import LazyImage from './../common/LazyImage';
 import ProductDetail from './ProductDetail';
@@ -34,27 +38,13 @@ const ProductItem = ({ product }) => {
         overflow='hidden'
         variant='outline'
         cursor='pointer'>
-        <AspectRatio width={{ base: '100%', sm: '200px' }} ratio={1}>
-          <LazyImage src={mainImage} alt={name} width='100%' />
-        </AspectRatio>
+        <ThumbNailImage product={product} />
 
         <Stack flex='1'>
           <CardBody position='relative'>
-            <Text
-              decoration='underline'
-              position='absolute'
-              top={5}
-              right={6}
-              minW={10}
-              textAlign='center'>
-              {idx}
-            </Text>
-            <Text fontSize='md' noOfLines={1} pr={10}>
-              {name}
-            </Text>
-            <Text fontWeight='bold' fontSize='md' py='2'>
-              {price.toLocaleString()}원
-            </Text>
+            <ProductNumber idx={idx} />
+            <ProductName name={name} />
+            <ProductPrice price={price} />
             <Tag>{spaceCategory}</Tag>
           </CardBody>
